// src/app/modules/Reservas/agency-management/models/agency.model.ts
export interface Agency {
  MR01_CodAgencia: string;
  MR01_Ruc: string;
  MR01_NomAgencia: string;
  MR01_Direccion: string;
  MR01_Ciudad: string;
  MR01_Pais: string;
  MR01_Primario: number;
  MR01_Mercado: string;
  MR01_Contacto: string;
  MR01_Telefono1: string;
  MR01_Telefono2: string;
  MR01_Fax1: string;
  MR01_Email: string;
  MR01_CodHabita: string;
  MR01_Numhabita: number;
  MR01_CodReserva: string;
  MR01_Correlativo: number;
  MR01_PorDescu: number;
  MR01_Activo: number;
  MR01_Operador: string;
  MR01_CodCliente: unknown;
  MR01_CodCHM: unknown;
  MR01_Prepago: unknown;
}